# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.ErrorResponse:
    properties:
      details:
        description: Additional error details
        example: Email field is required
        type: string
      error:
        description: HTTP error status text
        example: Bad Request
        type: string
      message:
        description: Human-readable error message
        example: Invalid request parameters
        type: string
    type: object
...
  models.OAuthLoginResponse:
    properties:
      access_token:
        description: JWT access token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      is_new_user:
        description: Whether this is a newly created user
        example: false
        type: boolean
      refresh_token:
        description: Refresh token for obtaining new access tokens
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Authenticated user information
    type: object
...
  models.Permission:
    properties:
      action:
        description: Action allowed on the resource (e.g., "create", "read", "update",
          "delete")
        example: delete
        type: string
      created_at:
        description: Timestamp when permission was created
        example: "2024-01-15T10:30:00Z"
        type: string
      description:
        description: Human-readable description of the permission
        example: Allows deleting users from the system
        type: string
      id:
        description: Permission unique identifier
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        description: Permission name (e.g., "users.delete", "api_keys.view")
        example: users.delete
        type: string
      resource:
        description: Resource the permission applies to (e.g., "users", "api_keys")
        example: users
        type: string
    required:
    - action
    - name
    - resource
    type: object
...
  models.Role:
    properties:
      created_at:
        description: Timestamp when role was created
        example: "2024-01-15T10:30:00Z"
        type: string
      description:
        description: Role description
        example: Full system access with all permissions
        type: string
      display_name:
        description: Human-readable display name
        example: Administrator
        type: string
      id:
        description: Role unique identifier
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_system_role:
        description: Whether this is a system-defined role (cannot be deleted)
        example: true
        type: boolean
      name:
        description: System name for the role
        example: admin
        type: string
      permissions:
        description: Many-to-many relation with Permission
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        description: Timestamp when role was last updated
        example: "2024-01-15T10:30:00Z"
        type: string
    required:
    - display_name
    - name
    type: object
...
  models.TelegramAuthData:
    properties:
      auth_date:
        description: Authentication timestamp
        example: 1234567890
        type: integer
      first_name:
        description: User's first name
        example: John
        type: string
      hash:
        description: Hash for verification
        example: abc123def456...
        type: string
      id:
        description: Telegram user ID
        example: 123456789
        type: integer
      last_name:
        description: User's last name (optional)
        example: Doe
        type: string
      photo_url:
        description: Profile photo URL (optional)
        example: https://t.me/i/userpic/123.jpg
        type: string
      username:
        description: Username without @ (optional)
        example: johndoe
        type: string
    type: object
...
  models.User:
    properties:
      account_type:
        description: 'Account type: "human" or "service"'
        example: human
        type: string
      created_at:
        description: Timestamp when user was created
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        description: User's email address
        example: user@example.com
        type: string
      email_verified:
        description: Whether email has been verified
        example: true
        type: boolean
      email_verified_at:
        description: Timestamp when email was verified
        example: "2024-01-15T10:30:00Z"
        type: string
      full_name:
        description: User's full name
        example: John Doe
        type: string
      id:
        description: Unique user identifier
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      is_active:
        description: Whether the account is active
        example: true
        type: boolean
      phone:
        description: User's phone number (optional)
        example: "+1234567890"
        type: string
      phone_verified:
        description: Whether phone number has been verified
        example: false
        type: boolean
      profile_picture_url:
        description: URL to user's profile picture
        example: https://example.com/avatars/user.jpg
        type: string
      roles:
        description: Relations
        items:
          $ref: '#/definitions/models.Role'
        type: array
      totp_enabled:
        description: Whether TOTP 2FA is enabled
        example: false
        type: boolean
      totp_enabled_at:
        description: Timestamp when TOTP 2FA was enabled
        example: "2024-01-15T10:30:00Z"
        type: string
      updated_at:
        description: Timestamp when user was last updated
        example: "2024-01-15T10:30:00Z"
        type: string
      username:
        description: Unique username for the user
        example: johndoe
        type: string
    type: object
...
host: localhost:8811
info:
  contact:
    email: maksbalashov@gmail.com
    name: API Support
  description: |-
    Centralized authentication and authorization system for microservices ecosystem.

    ## Overview
    Auth Gateway provides a complete authentication and authorization solution including:
    - User registration and authentication (email/password, OAuth, passwordless)
    - Two-factor authentication (TOTP)
    - API key management for service-to-service communication
    - Role-based access control (RBAC)
    - Session management
    - Webhooks for event notifications

    ## Authentication
    Most endpoints require authentication via JWT Bearer token or API key.

    ### JWT Authentication
    Include the access token in the Authorization header: `Authorization: Bearer {token}`

    ### API Key Authentication
    Include the API key in the X-API-Key header: `X-API-Key: agw_{key}`
  title: Auth Gateway API
  version: "1.0"
paths:
...
  /api/auth/{provider}:
    get:
      description: Redirect to OAuth provider for authentication (Google, Yandex,
        GitHub, Instagram, Telegram)
      parameters:
      - description: OAuth provider
        enum:
        - google
        - yandex
        - github
        - instagram
        - telegram
        in: path
        name: provider
        required: true
        type: string
      responses:
        "302":
          description: Redirect to OAuth provider
          schema:
            type: string
        "400":
          description: Invalid provider
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Initiate OAuth login
      tags:
      - OAuth
  /api/auth/{provider}/callback:
    get:
      description: Process OAuth callback from provider, create or login user, and
        redirect with tokens
      parameters:
      - description: OAuth provider
        enum:
        - google
        - yandex
        - github
        - instagram
        in: path
        name: provider
        required: true
        type: string
      - description: Authorization code from OAuth provider
        in: query
        name: code
        required: true
        type: string
      - description: CSRF state parameter
        in: query
        name: state
        required: true
        type: string
      - description: 'Response type: ''json'' for JSON response, otherwise redirect'
        enum:
        - json
        in: query
        name: response_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JSON response when response_type=json
          schema:
            $ref: '#/definitions/models.OAuthLoginResponse'
        "302":
          description: Redirect to frontend with tokens
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Handle OAuth callback
      tags:
      - OAuth
...
  /api/auth/telegram/callback:
    post:
      consumes:
      - application/json
      description: Process Telegram widget authentication data and return tokens
      parameters:
      - description: Telegram widget auth data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.TelegramAuthData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.OAuthLoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Invalid authentication data
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Handle Telegram auth callback
      tags:
      - OAuth
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: 'API Key for service-to-service authentication. Format: "agw_{key}"'
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token for user authentication. Format: "Bearer {access_token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"